requirements.txt
openai 
python-dotenv
python-docx